Separate swaping and cursor unlocking into separate scripts. Update scene
